Created Home View Model and HomeView is Refactored.
