+++ FitnessApp/FitnessAppApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct FitnessAppApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            FitnessTabView()
         }
     }
 }
