+++ FitnessApp/Home/Models/WorkoutModel.swift
@@ -1,0 +1,19 @@
+//
+//  WorkoutModel.swift
+//  FitnessApp
+//
+//  Created by Nikitha Tata on 30/07/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct WorkOut {
+    let id: Int
+    let title: String
+    let image: String
+    let duration: String
+    let date: String
+    let calories: String
+    let tintColor: Color
+}
